requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudscraper==1.2.48
colorama==0.4.4
commonmark==0.9.1
decorator==4.4.2
greenlet==1.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
moviepy==1.0.3
msgpack==1.0.2
mypy-extensions==0.4.3
neovim==0.3.1
numpy==1.20.0
opencv-python==4.5.1.48
parso==0.8.1
pathspec==0.8.1
Pillow==8.1.0
PixivPy==3.5.11
proglog==0.1.9
pydub==0.24.1
Pygments==2.7.4
pylint==2.6.0
pynvim==0.4.2
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
rich==9.10.0
six==1.15.0
toml==0.10.2
tqdm==4.56.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
yapf==0.30.0
